refactor(desktop): split AgentWorkspace into domain modules
Move the coding-agents workspace components out of the 1661-line
AgentWorkspace.tsx into focused modules with zero behavior change
(code moved verbatim, no logic edits):

- AgentWorkspacePanels.tsx: capabilityEnabled, InspectorEmptyState,
  NotificationPreferencesPanel, ProviderList, AttentionThreadList
- AgentComposer.tsx: mergeAttachments, mergeComposerSeed,
  clearComposerLaunchContext, ComposerSeed, AgentComposer
- AgentThreadLists.tsx: ThreadList, CreatedThreadHandleList
- AgentReviewDiff.tsx: review status/diff-line helpers, ReviewDiffLines
- AgentReviewPanel.tsx: ReviewList, ReviewSnapshotPanel,
  ReviewFileBrowserPanel, FileContentPanel, reviewHunkFollowUpDraft

AgentWorkspace.tsx keeps only the default-export composition root and
re-exports mergeAttachments, mergeComposerSeed, clearComposerLaunchContext
so existing imports keep working. Full desktop suite (1015 tests) and
desktop typecheck pass unchanged.
